refactor update & delete tryout
