+++ cms/src/components/Tryout/ListTryout.tsx
@@ -1,16 +1,19 @@
 import React, { useEffect, useState } from 'react';
-import { Table, Tag, Typography, DatePicker, Button, message, Modal, Form, Input, InputNumber } from 'antd';
+import { Table, Tag, Typography, DatePicker, Button, message, Input } from 'antd';
 import type { TableProps } from 'antd';
 import { CopyOutlined, DeleteOutlined, EditFilled, PlusOutlined } from '@ant-design/icons';
 import { useNavigate } from 'react-router-dom';
 import useFetchList from '../../hooks/useFetchList';
-import { TryoutProps, UpdateTryoutRequest } from '../../types/tryout.type';
+import { TryoutProps } from '../../types/tryout.type';
 import dayjs from 'dayjs';
-import { apiCreateTryout, apiDeleteTryout, apiUpdateTryout } from '../../api/tryout';
 import type { DatePickerProps, GetProps } from 'antd';
 import { getErrorMessage } from '../../helpers/errorHandler';
 import copy from 'copy-to-clipboard';
 import SwitchButton from '../Ui/SwitchButton';
+import ModalCreateTryout from './ModalCreateTryout';
+import ModalUpdateTryout from './ModalUpdateTryout';
+import ModalDeleteTryout from './ModalDeleteTryout';
+import { apiUpdateTryout } from '../../api/tryout';
 type RangePickerProps = GetProps<typeof DatePicker.RangePicker>;
 
 const { Title } = Typography;
@@ -20,7 +23,6 @@ const { Text, Link } = Typography;
 const ListTryout = () => {
   const navigate = useNavigate();
   const [showModal, setShowModal] = useState(false);
-  const [isPublished, setIsPublished] = useState(false);
   const [showModalUpdate, setShowModalUpdate] = useState({
     status: false,
     data: {} as any,
@@ -33,57 +35,10 @@ const ListTryout = () => {
     endpoint: 'tryout',
   });
 
-  const [form] = Form.useForm();
-  const [updateForm] = Form.useForm();
-
   useEffect(() => {
     fetchList();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-
-  useEffect(() => {
-    if (showModalUpdate.status) {
-      updateForm.setFieldsValue({
-        ...showModalUpdate.data,
-        start_time: dayjs(showModalUpdate.data.start_time),
-        end_time: dayjs(showModalUpdate.data.end_time),
-      });
-      setIsPublished(showModalUpdate?.data?.is_published);
-    }
-  }, [showModalUpdate, updateForm]);
-
-  const handleUpdate = async () => {
-    try {
-      const values = await updateForm.validateFields();
-      console.log(values);
-
-      const data: UpdateTryoutRequest = {
-        ...values,
-      };
-      console.log(data);
-
-      const res = await apiUpdateTryout(showModalUpdate.data.tryout_id, data);
-      console.log(res);
-
-      if (res) {
-        setShowModalUpdate({
-          status: false,
-          data: {},
-        });
-      }
-      fetchList();
-      message.success('Success Update');
-    } catch (err) {
-      message.error('Failed Update');
-    }
-  };
-
-  const handleDelete = async () => {
-    await apiDeleteTryout(showModalDelete.tryoutId);
-    fetchList();
-    setShowModalDelete({ status: false, tryoutId: '' });
-    message.success('Success Delete');
-  };
 
   const columns: TableProps<TryoutProps>['columns'] = [
     {
@@ -122,6 +77,20 @@ const ListTryout = () => {
       key: 'status',
       dataIndex: 'status',
       render: (_, { start_time, end_time }) => <Tag color={statusColor(checkStatus(start_time, end_time))}>{checkStatus(start_time, end_time)}</Tag>,
+    },
+    {
+      title: 'Published',
+      key: 'is_published',
+      dataIndex: 'is_published',
+      render: (_, { tryout_id, is_published }) => (
+        <SwitchButton
+          defaultChecked={is_published}
+          onChange={(checked) => {
+            handlePublished(tryout_id, checked);
+            console.log('Published: ', checked);
+          }}
+        />
+      ),
     },
     {
       title: 'Token',
@@ -190,22 +159,20 @@ const ListTryout = () => {
     }
   };
 
-  const handleCreate = async (data: any) => {
-    console.log(data);
-    const newData = {
-      ...data,
-      is_published: isPublished,
-    };
-    console.log(newData);
+  const handlePublished = async (tryoutId: string, checked: boolean) => {
+    try {
+      const res = await apiUpdateTryout(tryoutId, { is_published: checked });
+      console.log(res?.data);
 
-    const res = await apiCreateTryout(newData);
-    console.log(res);
-    if (res) {
-      setShowModal(false);
-      fetchList();
-      message.success('Create Tryout Success');
+      if (res) {
+        fetchList();
+        message.success('Success Update Published Status');
+      }
+    } catch (err) {
+      message.error('Failed Published');
     }
   };
+
   const onFinishFailed = (errorInfo: any) => {
     message.error(getErrorMessage(errorInfo));
   };
@@ -242,212 +209,30 @@ const ListTryout = () => {
         dataSource={tryoutData}
       />
 
-      <Modal
-        open={showModal}
-        onCancel={() => setShowModal(false)}
-        footer={false}
-      >
-        <Title
-          level={3}
-          style={{ fontWeight: 'bold' }}
-        >
-          Create Tryout
-        </Title>
-        <Form
-          name="createTryout"
-          onFinish={handleCreate}
-          onFinishFailed={onFinishFailed}
-          initialValues={{ is_published: false }}
-          layout="vertical"
-          form={form}
-        >
-          <Form.Item
-            label="Tryout Name"
-            name="title"
-            rules={[{ required: true }]}
-          >
-            <Input />
-          </Form.Item>
+      <ModalCreateTryout
+        showModal={showModal}
+        setShowModal={setShowModal}
+        onFinishFailed={onFinishFailed}
+        onChange={onChange}
+        onOk={onOk}
+        fetchList={fetchList}
+      />
 
-          <Form.Item
-            label="Duration (Minutes)"
-            name="duration"
-            rules={[{ required: true }]}
-          >
-            <InputNumber />
-          </Form.Item>
+      <ModalUpdateTryout
+        showModalUpdate={showModalUpdate}
+        setShowModalUpdate={setShowModalUpdate}
+        onFinishFailed={onFinishFailed}
+        onChange={onChange}
+        onOk={onOk}
+        fetchList={fetchList}
+      />
 
-          <Form.Item
-            label="Start Time"
-            name="start_time"
-            rules={[{ required: true }]}
-          >
-            <DatePicker
-              showTime
-              onChange={onChange}
-              onOk={onOk}
-            />
-          </Form.Item>
-
-          <Form.Item
-            label="End Time"
-            name="end_time"
-            rules={[{ required: true }]}
-          >
-            <DatePicker
-              showTime
-              onChange={onChange}
-              onOk={onOk}
-            />
-          </Form.Item>
-
-          <Form.Item
-            label="Published"
-            name="is_published"
-            // valuePropName="checked"
-          >
-            <SwitchButton
-              defaultChecked={isPublished}
-              onChange={(checked) => {
-                setIsPublished(checked);
-                console.log('Published: ', checked);
-              }}
-            />
-          </Form.Item>
-
-          <Form.Item wrapperCol={{ offset: 8, span: 16 }}>
-            <Button
-              type="primary"
-              htmlType="submit"
-            >
-              Submit
-            </Button>
-          </Form.Item>
-        </Form>
-      </Modal>
-
-      <Modal
-        open={showModalUpdate.status}
-        onCancel={() => setShowModalUpdate({ data: {}, status: false })}
-        footer={false}
-      >
-        <Title
-          level={3}
-          style={{ fontWeight: 'bold' }}
-        >
-          Update Tryout
-        </Title>
-        <Form
-          name="updateTryout"
-          onFinish={handleUpdate}
-          onFinishFailed={onFinishFailed}
-          // initialValues={{ is_published: false }}
-          layout="vertical"
-          form={updateForm}
-        >
-          <Form.Item
-            label="Tryout Name"
-            name="title"
-            rules={[{ required: true }]}
-          >
-            <Input defaultValue={showModalUpdate?.data?.title} />
-          </Form.Item>
-
-          <Form.Item
-            label="Duration (Minutes)"
-            name="duration"
-            rules={[{ required: true }]}
-          >
-            <InputNumber defaultValue={showModalUpdate?.data?.duration} />
-          </Form.Item>
-
-          <Form.Item
-            label="Start Time"
-            name="start_time"
-            rules={[{ required: true }]}
-          >
-            <DatePicker
-              showTime
-              onChange={onChange}
-              onOk={onOk}
-              defaultValue={dayjs(showModalUpdate?.data?.start_time)}
-            />
-          </Form.Item>
-
-          <Form.Item
-            label="End Time"
-            name="end_time"
-            rules={[{ required: true }]}
-          >
-            <DatePicker
-              showTime
-              onChange={onChange}
-              onOk={onOk}
-              defaultValue={dayjs(showModalUpdate?.data?.end_time)}
-            />
-          </Form.Item>
-
-          <Form.Item
-            label="Published"
-            name="is_published"
-            // valuePropName="checked"
-          >
-            <SwitchButton
-              defaultChecked={showModalUpdate?.data?.is_published}
-              onChange={(checked) => {
-                setIsPublished(checked);
-                console.log('Published: ', checked);
-              }}
-            />
-          </Form.Item>
-
-          <Form.Item wrapperCol={{ offset: 8, span: 16 }}>
-            <Button
-              type="primary"
-              htmlType="submit"
-            >
-              Submit
-            </Button>
-          </Form.Item>
-        </Form>
-      </Modal>
-      <Modal
-        open={showModalDelete.status}
-        onCancel={() => setShowModalDelete({ status: false, tryoutId: '' })}
-        footer={false}
-      >
-        <Title
-          level={4}
-          style={{ fontWeight: 'semibold' }}
-        >
-          Are you sure to delete this tryout?
-        </Title>
-        <Form
-          name="deleteTryout"
-          onFinish={handleDelete}
-          onFinishFailed={onFinishFailed}
-          // initialValues={{ is_published: false }}
-          layout="vertical"
-        >
-          <Form.Item style={{ display: 'flex', alignItems: 'center', gap: '20px', width: '100%', justifyContent: 'end' }}>
-            <Button
-              danger
-              type="default"
-              htmlType="button"
-              onClick={() => setShowModalDelete({ status: false, tryoutId: '' })}
-            >
-              Cancel
-            </Button>
-            <Button
-              type="primary"
-              htmlType="submit"
-              style={{ marginLeft: '10px' }}
-            >
-              Delete
-            </Button>
-          </Form.Item>
-        </Form>
-      </Modal>
+      <ModalDeleteTryout
+        showModalDelete={showModalDelete}
+        setShowModalDelete={setShowModalDelete}
+        onFinishFailed={onFinishFailed}
+        fetchList={fetchList}
+      />
     </div>
   );
 };
